Expand the content lib
